Rename Service to AbstractService
It is more appropriate as the class must be extended to work correctly.
