+++ packages/js-toolkit/services/index.ts
@@ -1,3 +1,4 @@
+export * from './AbstractService.js';
 export * from './DragService.js';
 export * from './KeyService.js';
 export * from './LoadService.js';
@@ -5,4 +6,3 @@ export * from './PointerService.js';
 export * from './RafService.js';
 export * from './ResizeService.js';
 export * from './ScrollService.js';
-export * from './Service.js';
